# Java: Tuple serialization/deserialization uses fewer allocations