--- a/bindings/java/src/main/com/apple/foundationdb/tuple/IterableComparator.java
+++ b/bindings/java/src/main/com/apple/foundationdb/tuple/IterableComparator.java
@@ -34,7 +34,7 @@ import java.util.Iterator;
  *    tuple1.compareTo(tuple2)
  *      == new IterableComparator().compare(tuple1, tuple2)
  *      == new IterableComparator().compare(tuple1.getItems(), tuple2.getItems()),
- *      == ByteArrayUtil.compareUnsigned(tuple1.packInternal(), tuple2.packInternal())}
+ *      == ByteArrayUtil.compareUnsigned(tuple1.pack(), tuple2.pack())}
  * </pre>
  *
  * <p>