--- a/bindings/java/src/main/com/apple/foundationdb/tuple/Versionstamp.java
+++ b/bindings/java/src/main/com/apple/foundationdb/tuple/Versionstamp.java
@@ -94,8 +94,8 @@ public class Versionstamp implements Comparable<Versionstamp> {
 	private static final byte[] UNSET_TRANSACTION_VERSION = {(byte)0xff, (byte)0xff, (byte)0xff, (byte)0xff, (byte)0xff,
 	                                                         (byte)0xff, (byte)0xff, (byte)0xff, (byte)0xff, (byte)0xff};
 
-	private boolean complete;
-	private byte[] versionBytes;
+	private final boolean complete;
+	private final byte[] versionBytes;
 
 	/**
 	 * From a byte array, unpack the user version starting at the given position.