… bindings/java/CMakeLists.txt
… bindings/java/src/test/com/apple/foundationdb/test/TuplePerformanceTest.java
… bindings/java/src/test/com/apple/foundationdb/test/TupleTest.java
… documentation/sphinx/source/release-notes.rst