This makes a few improvements to the Java Tuple packing and unpacking logic. The main thing is it now calculates the size of the tuple it is about to pack before allocating one array and then packing things into it.

This also:

* Uses `long`s instead of `BigInteger`s if possible
* Uses stack-allocated primitives for `float` and `double` serialization/deserialization
* Avoids copying byte array and string representations for comparison
* Memoizes the serialized form for fast reserialization
* Adds a method for the user to pack a tuple into a `ByteBuffer`
* Fixes two bugs related to malformed tuples (while I was here)

To give a feel for the improvement here, here's a table with the performance results packing random tuples:

|     | Master | With PR | Improvement |
|---|--------|----------| --------- |
| Bytes per Tuple | 141 B |  141 B |  N/A |
| Pack time | 2.52 µs | 1.39 µs | 1.81 x |
| Unpack time | 1.33 µs | 0.85 µs | 1.56 x |

Here are the numbers specifically for random tuples consisting only of `long`s to test no longer using `BigInteger`s:

|     | Master | With PR | Improvement |
|---|--------|----------| --------- |
| Bytes per Tuple | 47.5 B |  47.5 B |  N/A |
| Pack time | 1.17 µs | 0.424 µs | 2.76 x |
| Unpack time | 0.720 µs | 0.321 µs | 2.24 x |

Here are numbers for random tuples consisting only of floating-point numbers (both floats and doubles):

|     | Master | With PR | Improvement |
|---|--------|----------| --------- |
| Bytes per Tuple | 66.5 B |  66.5 B |  N/A |
| Pack time | 0.435 µs | 0.282 µs | 1.54 x |
| Unpack time | 0.299 µs | 0.244 µs | 1.23 x |

And here are numbers for random tuples consisting only of byte arrays and strings (or "string like") values:

|     | Master | With PR | Improvement |
|---|--------|----------| --------- |
| Bytes per Tuple | 145 B |  145 B |  N/A |
| Pack time | 3.21 µs | 1.37 µs | 2.34 x |
| Unpack time | 2.40 µs | 1.01 µs | 2.38 x |

I tried to add tests to cover what was changed. They mainly got placed in `TupleTest` with performance testing changes added to `TuplePerformanceTest`. I will attempt to run some randomized bindingtester tests as well.

This resolves #1206. It also fixes #671 and it fixes #672.