@ajbeamon 2019-03-01
I assume this does fine in the binding tester?