@alecgrieser 2019-03-02
> I assume this does fine in the binding tester?

I ran about 100,000 bindingtester runs, and they were all clean. However, not all of those were Java tests--something like 2/7 of the runs.